chore: refactor message components
