+++ web/screens/Thread/ThreadCenterPanel/ChatItem/index.tsx
@@ -1,6 +1,7 @@
-import React, { forwardRef, useEffect, useState } from 'react'
+import React, { forwardRef, useEffect, useMemo, useState } from 'react'
 
 import {
+  ContentType,
   events,
   MessageEvent,
   MessageStatus,
@@ -8,9 +9,20 @@ import {
   ThreadMessage,
 } from '@janhq/core'
 
+import { Tooltip } from '@janhq/joi'
+
+import { FolderOpenIcon } from 'lucide-react'
+
 import ErrorMessage from '@/containers/ErrorMessage'
 
-import SimpleTextMessage from '../SimpleTextMessage'
+import { usePath } from '@/hooks/usePath'
+
+import { toGibibytes } from '@/utils/converter'
+import { openFileTitle } from '@/utils/titleUtils'
+
+import Icon from '../FileUploadPreview/Icon'
+import TextMessage from '../TextMessage'
+import { RelativeImage } from '../TextMessage/RelativeImage'
 
 type Ref = HTMLDivElement
 
@@ -27,6 +39,13 @@ const ChatItem = forwardRef<Ref, Props>((message, ref) => {
       ? message
       : undefined
   )
+  const messageType = useMemo(() => content[0]?.type ?? '', [content])
+
+  const annotation = useMemo(
+    () => content[0]?.text?.annotations[0] ?? '',
+    [content]
+  )
+  const { onViewFile, onViewFileContainer } = usePath()
 
   function onMessageUpdate(data: ThreadMessage) {
     if (data.id === message.id) {
@@ -54,7 +73,60 @@ const ChatItem = forwardRef<Ref, Props>((message, ref) => {
     <>
       {status !== MessageStatus.Error && content?.length > 0 && (
         <div ref={ref} className="relative">
-          <SimpleTextMessage {...message} content={content} status={status} />
+          {messageType === ContentType.Image && (
+            <div className="group/image relative mb-2 inline-flex cursor-pointer overflow-hidden rounded-xl">
+              <div className="left-0 top-0 z-20 h-full w-full group-hover/image:inline-block">
+                <RelativeImage
+                  src={annotation}
+                  id={message.id}
+                  onClick={() => onViewFile(annotation)}
+                />
+              </div>
+              <Tooltip
+                trigger={
+                  <div
+                    className="absolute right-2 top-2 z-20 hidden h-8 w-8 cursor-pointer items-center justify-center rounded-md bg-[hsla(var(--app-bg))] group-hover/image:flex"
+                    onClick={onViewFileContainer}
+                  >
+                    <FolderOpenIcon size={20} />
+                  </div>
+                }
+                content={<span>{openFileTitle()}</span>}
+              />
+            </div>
+          )}
+
+          {messageType === ContentType.Pdf && (
+            <div className="group/file bg-secondary relative mb-2 inline-flex w-60 cursor-pointer gap-x-3 overflow-hidden rounded-lg p-4">
+              <div
+                className="absolute left-0 top-0 z-20 hidden h-full w-full bg-black/20 backdrop-blur-sm group-hover/file:inline-block"
+                onClick={() => onViewFile(`${message.id}.${messageType}`)}
+              />
+              <Tooltip
+                trigger={
+                  <div
+                    className="absolute right-2 top-2 z-20 hidden h-8 w-8 cursor-pointer items-center justify-center rounded-md bg-[hsla(var(--app-bg))] group-hover/file:flex"
+                    onClick={onViewFileContainer}
+                  >
+                    <FolderOpenIcon size={20} />
+                  </div>
+                }
+                content={<span>{openFileTitle()}</span>}
+              />
+              <Icon type={content[0].type} />
+              <div className="w-full">
+                <h6 className="line-clamp-1 w-4/5 font-medium">
+                  {content[0].text.name?.replaceAll(/[-._]/g, ' ')}
+                </h6>
+                <p className="text-[hsla(var(--text-secondary)]">
+                  {toGibibytes(Number(content[0].text.size))}
+                </p>
+              </div>
+            </div>
+          )}
+          {messageType === ContentType.Text && (
+            <TextMessage {...message} content={content} status={status} />
+          )}
         </div>
       )}
       {errorMessage && !message.loadModelError && (
